+++ httprunner/core/testcase/step/runapi/request.py
@@ -6,6 +6,7 @@ from typing import (
 from httprunner.core.testcase.beforestep.param import ParametrizeStepMixin
 from httprunner.core.testcase.beforestep.retry import RetryStepMixin
 from httprunner.core.testcase.beforestep.skip import SkipStepMixin
+from httprunner.core.testcase.beforestep.timer import TimerMixin
 from httprunner.core.testcase.beforestep.variable import StepVariableMixin
 from httprunner.core.testcase.config import Config  # noqa
 from httprunner.core.testcase.step.hook.setup import SetupHookMixin
@@ -31,6 +32,7 @@ class HttpRunnerRequest(
     RetryStepMixin,
     StepVariableMixin,
     SetupHookMixin,
+    TimerMixin,
     RequestWithOptionalArgs,
 ):
     """