+++ httprunner/runner.py
@@ -26,6 +26,7 @@ from httprunner.core.runner.retry import (
 )
 from httprunner.core.runner.skip_step import is_skip_step
 from httprunner.core.runner.step_shell_variables import get_step_shell_variables
+from httprunner.core.runner.timer import display_delay_in_step_name
 from httprunner.core.runner.update_form import update_form
 from httprunner.core.runner.update_json import update_json
 from httprunner.core.runner.with_resource import evaluate_resources
@@ -539,6 +540,10 @@ class HttpRunner(object):
         """Core function for running step (maybe a request or referenced testcase)."""
         self.__resolve_step_variables(step)
 
+        if step.pre_delay_seconds:
+            logger.info(f"Sleep Before: {step.pre_delay_seconds} seconds")
+            time.sleep(step.pre_delay_seconds)
+
         try:
             logger.info(f"run step begin: {step.name} >>>>>>")
 
@@ -556,6 +561,10 @@ class HttpRunner(object):
             raise e
         finally:
             logger.info(f"run step end: {step.name} <<<<<<\n")
+
+        if step.post_delay_seconds:
+            logger.info(f"Sleep After: {step.post_delay_seconds} seconds")
+            time.sleep(step.post_delay_seconds)
 
     def __parse_step_name(self, step: TStep) -> str:
         """Parse step name with step context variables and variables defined by step self."""
@@ -589,11 +598,13 @@ class HttpRunner(object):
             # so the original step should stay untouched.
             step_copy = step.model_copy(deep=True)
             self.__resolve_step_variables(step_copy)
+            display_delay_in_step_name(step_copy, self.__project_meta.functions)
             return parse_data(
                 step_copy.name, step_copy.variables, self.__project_meta.functions
             )
         else:
             self.__resolve_step_variables(step)
+            display_delay_in_step_name(step, self.__project_meta.functions)
             return parse_data(step.name, step.variables, self.__project_meta.functions)
 
     def __run_step(self, step: TStep) -> NoReturn: